extract formatElapsed function
