+++ lib/index.js
@@ -1,3 +1,5 @@
+const formatElapsed = require('./util').formatElapsed
+
 module.exports = Clocker
 
 var path = require('path')
@@ -48,14 +50,6 @@ Clocker.getDataObject = function (row) {
   }
 
   return obj
-}
-
-Clocker.formatElapsed = function (elapsed) {
-  var n = elapsed
-  var hh = pad(Math.floor(n / 60 / 60), 2)
-  var mm = pad(Math.floor(n / 60 % 60), 2)
-  var ss = pad(Math.floor(n % 60), 2)
-  return [ hh, mm, ss ].join(':')
 }
 
 Clocker.prototype.close = function (cb) {
@@ -176,7 +170,7 @@ Clocker.prototype.status = function (id, cb) {
     if (!row.end) {
       var elapsed = (new Date()) - started
       var seconds = elapsed / 1000
-      return cb(null, 'elapsed time: ' + Clocker.formatElapsed(seconds))
+      return cb(null, 'elapsed time: ' + formatElapsed(seconds))
     }
 
     return cb(null, 'stopped')
@@ -480,10 +474,6 @@ function addData (old, update) {
   }
 
   return old
-}
-
-function pad (s, len) {
-  return Array(Math.max(0, len - String(s).length + 1)).join('0') + s
 }
 
 function getKey (x) {
