+++ lib/util.js
@@ -1,7 +1,3 @@
-module.exports = {}
-module.exports.getDate = getDate
-module.exports.isUnixTimestamp = isUnixTimestamp
-
 var parseTime = require('parse-messy-time')
 
 function getDate (expr) {
@@ -25,4 +21,22 @@ function getDate (expr) {
 
 function isUnixTimestamp (expr) {
   return /^1[0-9]{9}$/.test(expr)
+}
+
+function formatElapsed (elapsed) {
+  var n = elapsed
+  var hh = pad(Math.floor(n / 60 / 60), 2)
+  var mm = pad(Math.floor(n / 60 % 60), 2)
+  var ss = pad(Math.floor(n % 60), 2)
+  return [ hh, mm, ss ].join(':')
+}
+
+function pad (s, len) {
+  return Array(Math.max(0, len - String(s).length + 1)).join('0') + s
+}
+
+module.exports = {
+  getDate,
+  isUnixTimestamp,
+  formatElapsed
 }
